# perf(nodejs): introduce pool into default rng