@broofa 2020-08-27
FWIW, results on my laptop...

Before:
```
uuidv1() x 1,586,023 ops/sec ±0.73% (91 runs sampled)
uuidv1() fill existing array x 8,903,577 ops/sec ±1.07% (89 runs sampled)
uuidv4() x 387,462 ops/sec ±3.11% (88 runs sampled)
uuidv4() fill existing array x 534,144 ops/sec ±2.71% (94 runs sampled)
```
After:
```
uuidv1() x 1,590,931 ops/sec ±0.55% (92 runs sampled)
uuidv1() fill existing array x 8,926,106 ops/sec ±0.56% (91 runs sampled)
uuidv4() x 745,416 ops/sec ±1.83% (82 runs sampled)
uuidv4() fill existing array x 1,685,734 ops/sec ±0.44% (96 runs sampled)
```