@puzpuzpuz 2020-08-28
@broofa I've addressed your comments and updated benchmark results in the PR description