@broofa 2020-08-29
@ctavan 'Not sure why bundlewatch test is hanging here.  I was able to get it to pass on my laptop so I'm merging, but I ran into some sort of "call stack exceeded" error initially when doing `npm run bundlewatch`.  Ended up `rm -r node_modules`'ing and reinstalling to get it to work.