@ctavan 2020-08-29
@broofa bundlewatch needs a token which is not available in CI runs from forks, only from branches within this repo... so PRs from forks will never report bundlewatch results. I haven’t found a workaround so far.