@simlu 2020-10-06
Please consider the regression introduced with this change.

https://github.com/uuidjs/uuid/issues/525

Looking forward to having a good discussion